+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## [Alpha 1.3.2](https://github.com/Nevah5/wordlebot/releases/tag/a1.3.2) - 07.02.2022
 ### Changed
 - Keyboard layout in game embeds
+- Keyboard styling in game embeds with lines on top and under
 
 ### Fixed
 - Keyboard bug where letters not avaiable